lib: Factor out definition of ofputil_protocol.
Compiling ofp-util.h with a C++ compiler fails because
meta-flow.h (included from ofp-util.h) uses this enum
prior to its definition in ofp-util.h. This is mainly
due to differences in enum handling between C and C++.
This change moves the definition of the enum to its
own file.

Signed-off-by: Amit Bose <bose@noironetworks.com>