Try to keep igmp table consistent with ovs mcast table.
igmp table entries can be deleted or added outside of the ovs api.

For the add case this means the entry was added via an ip link api or
some other mechanism. In this case add an entry to the hash table and
ignore the error.

For the delete case the igmp code would fail if entry does not exist.
This could happen if the underlying device was deleted or a change in
the route. since ovs will not know about this event our hash table
may have an entry that no longer exists. In this case simply delete
the entry.

The consistency is only checked for during an add or delete. So a get
could still return entries that may be out of sync with what the igmp
code has.

Perhaps in the long term we should get rid of our hash table and get
the list walking the igmp list.